backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
reportlab==4.0.7
emails==0.6.0
jinja2==3.1.2
python-dateutil==2.8.2